added 'transform' function and refactored code
