+++ gensim/sklearn_integration/__init__.py
@@ -16,5 +16,5 @@ from .sklearn_wrapper_gensim_lsimodel import SklLsiModel  # noqa: F401
 from .sklearn_wrapper_gensim_rpmodel import SklRpModel  # noqa: F401
 from .sklearn_wrapper_gensim_ldaseqmodel import SklLdaSeqModel  # noqa: F401
 from .sklearn_wrapper_gensim_w2vmodel import SklW2VModel  # noqa: F401
-from .sklearn_wrapper_gensim_d2vmodel import SklD2VModel  # noqa: F401
+from .d2vmodel import SklD2VModel  # noqa: F401
 from .sklearn_wrapper_gensim_atmodel import SklATModel  # noqa: F401
