+++ gensim/sklearn_integration/d2vmodel.py
@@ -9,14 +9,16 @@ Scikit learn interface for gensim for easy use of gensim with scikit-learn
 Follows scikit-learn API conventions
 """
 
+from numpy import integer
+from six import string_types, integer_types
 from sklearn.base import TransformerMixin, BaseEstimator
 from sklearn.exceptions import NotFittedError
 
 from gensim import models
-from gensim.sklearn_integration import BaseSklearnWrapper
+from gensim.sklearn_api import BaseTransformer
 
 
-class SklD2VModel(BaseSklearnWrapper, TransformerMixin, BaseEstimator):
+class D2VTransformer(BaseTransformer, TransformerMixin, BaseEstimator):
     """
     Base Doc2Vec module
     """
@@ -54,7 +56,7 @@ class SklD2VModel(BaseSklearnWrapper, TransformerMixin, BaseEstimator):
         """
         Set all parameters.
         """
-        super(SklD2VModel, self).set_params(**parameters)
+        super(D2VTransformer, self).set_params(**parameters)
         return self
 
     def fit(self, X, y=None):
@@ -68,11 +70,23 @@ class SklD2VModel(BaseSklearnWrapper, TransformerMixin, BaseEstimator):
             trim_rule=self.trim_rule, **self.other_params)
         return self
 
-    def transform(self, words):
+    def transform(self, docs):
         """
+        Return the vector representations for the input list of documents.
         """
         if self.gensim_model is None:
             raise NotFittedError("This model has not been fitted yet. Call 'fit' with appropriate arguments before using this method.")
 
+        # The input as array of array
+        check = lambda x: [x] if isinstance(x, string_types + integer_types + (integer,)) else x
+        docs = check(docs)
+        X = [[] for _ in range(0, len(docs))]
+
+        for k, v in enumerate(docs):
+            doc_vec = self.gensim_model[v]
+            X[k] = doc_vec
+
+        return np.reshape(np.array(X), (len(docs), self.gensim_model.vector_size))
+
     def partial_fit(self, X):
-        raise NotImplementedError("'partial_fit' has not been implemented for SklD2VModel")
+        raise NotImplementedError("'partial_fit' has not been implemented for D2VTransformer")
